+++ src/lib/data/queries.ts
@@ -59,3 +59,14 @@ export const getInstantiatedListByUserQueryDocument = graphql(`
     }
   }
 `);
+
+export const getInstantiateDetailByContractQueryDocument = graphql(`
+  query getInstantiateDetailByContractQueryDocument($contractAddress: String!) {
+    contracts_by_pk(address: $contractAddress) {
+      init_msg
+      transaction {
+        hash
+      }
+    }
+  }
+`);